Extract typeOf to it's own file
