+++ lib/sinon.js
@@ -16,6 +16,7 @@ var sinon = (function () {
     function loadDependencies(require, exports, module) {
         sinon = module.exports = require("./sinon/util/core");
         require("./sinon/extend");
+        require("./sinon/typeOf");
         require("./sinon/times_in_words");
         require("./sinon/spy");
         require("./sinon/call");
