+++ lib/sinon/match.js
@@ -1,4 +1,7 @@
-/* @depend util/core.js */
+/**
+ * @depend util/core.js
+ * @depend typeOf.js
+ */
 /*jslint eqeqeq: false, onevar: false, plusplus: false*/
 /*global module, require, sinon*/
 /**
