+++ lib/sinon/util/core.js
@@ -269,16 +269,6 @@
             });
         };
 
-        sinon.typeOf = function (value) {
-            if (value === null) {
-                return "null";
-            } else if (value === undefined) {
-                return "undefined";
-            }
-            var string = Object.prototype.toString.call(value);
-            return string.substring(8, string.length - 1).toLowerCase();
-        };
-
         sinon.createStubInstance = function (constructor) {
             if (typeof constructor !== "function") {
                 throw new TypeError("The constructor should be a function.");
