+++ test/sinon.html
@@ -35,6 +35,7 @@
     <script src="../lib/sinon/util/core.js"></script>
     <script src="../lib/sinon/extend.js"></script>
     <script src="../lib/sinon/times_in_words.js"></script>
+    <script src="../lib/sinon/typeOf.js"></script>
     <script src="../lib/sinon/match.js"></script>
     <script src="../lib/sinon/spy.js"></script>
     <script src="../lib/sinon/call.js"></script>
@@ -61,6 +62,7 @@
     <script src="sinon_test.js"></script>
     <script src="sinon/extend_test.js"></script>
     <script src="sinon/times_in_words_test.js"></script>
+    <script src="sinon/typeOf_test.js"></script>
     <script src="sinon/spy_test.js"></script>
     <script src="sinon/call_test.js"></script>
     <script src="sinon/stub_test.js"></script>
